Fixed bugs while rendering FlatList and changed to dark mode
I was trying to render FlatList with an array of arrays instead of an array of objects, so FlatList was getting an error. Fixed it by pushing the data to the array using a model. Did the same to Search screen and add some states to render the flatlist title after users click in button. Also add a clean button in Search, to clean all rendered data and return a blank screen again